# openapi_specification.yaml
openapi: 3.0.3
info:
  title: «Дәресханә» (Tatar OCR) — API Contract Specification
  version: 1.0.0
  description: |
    This document provides the definitive API contracts and architectural design for **«Дәресханә»**, built around two foundational principles:
    1. **Zero-Photo Transmission**: Camera images never leave the teacher's phone. All ArUco perspective rectifications, cell cropping, and TFLite neural network character recognitions happen 100% locally on the device in memory.
    2. **Backend Analytical Processing**: The phone uploads only lightweight structured verification results (1–2 KB JSON per sheet). The backend performs all heavy historical data aggregation, student learning curve analysis, class heatmaps, error categorizations, and gradebook exports.
servers:
  - url: /
    description: Base Server URL
components:
  securitySchemes:
    TeacherUUID:
      type: apiKey
      in: header
      name: X-Teacher-UUID
      description: Teacher UUID v4 generated upon first install
  parameters:
...
    HeaderTeacherUUID:
      name: X-Teacher-UUID
      in: header
      required: false
      description: Teacher UUID v4
      schema:
        type: string
        format: uuid
        example: "c56a4180-65aa-42ec-a945-5fd21dec0538"
  schemas:
...
    AssignmentMetadataResponse:
      type: object
      description: Metadata, title, grade level, topic tags, points.
      properties:
        assignment_id:
          type: string
          example: "TAT-2026-Q1"
        title:
          type: string
          example: "Татар теле. 7 сыйныф. Исем килешләре"
        grade_level:
          type: integer
          example: 7
        topic_tags:
          type: array
          items:
            type: string
          example: ["case_dative", "case_ablative"]
        points:
          type: number
          example: 8.0
    TemplateCellDimensions:
      type: object
      properties:
        width:
          type: number
          example: 10.0
        height:
          type: number
          example: 10.0
    TemplateGeometry:
      type: object
      properties:
        format:
          type: string
          example: "A4"
        corner_aruco_dict:
          type: string
          example: "DICT_4X4_50"
        corner_aruco_ids:
          type: array
          items:
            type: integer
          example: [0, 1, 2, 3]
        cell_dimensions_mm:
          $ref: '#/components/schemas/TemplateCellDimensions'
    ExpectedCell:
      type: object
      properties:
        index:
          type: integer
          example: 0
        char:
          type: string
          example: "К"
        unicode:
          type: string
          example: "U+041A"
        is_empty_allowed:
          type: boolean
          example: true
    OfflineBundleQuestion:
      type: object
      properties:
        question_number:
          type: integer
          example: 1
        marker_id:
          type: integer
          example: 11
        prompt:
          type: string
          example: "Куегыз сүзне юнәлеш килешендә: китап ->"
        topic_tag:
          type: string
          example: "case_dative"
        topic_name_tt:
          type: string
          example: "Юнәлеш килеше"
        cell_count:
          type: integer
          example: 8
        expected_answer:
          type: string
          example: "КИТАПКА"
        expected_cells:
          type: array
          items:
            $ref: '#/components/schemas/ExpectedCell'
    OfflineBundleVariant:
      type: object
      properties:
        variant_id:
          type: integer
          example: 1
        qr_signature:
          type: string
          example: "{\"tid\":\"TAT-2026-Q1\",\"var\":1,\"page\":1,\"tot\":1,\"n_q\":8}"
        template_geometry:
          $ref: '#/components/schemas/TemplateGeometry'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/OfflineBundleQuestion'
    OfflineBundleResponse:
      type: object
      properties:
        assignment_id:
          type: string
          example: "TAT-2026-Q1"
        title:
          type: string
          example: "Татар теле. 7 сыйныф. Исем килешләре"
        total_variants:
          type: integer
          example: 2
        variants:
          type: array
          items:
            $ref: '#/components/schemas/OfflineBundleVariant'
    TaskBankItem:
      type: object
      properties:
        task_id:
          type: string
          example: "tsk_8201"
        prompt_tt:
          type: string
          example: "Сүзгә күплек сан кушымчасын ялгагыз: дус ->"
        expected_answer:
          type: string
          example: "ДУСЛАР"
        cell_count:
          type: integer
          example: 6
        grade_level:
          type: integer
          example: 5
        topic_tag:
          type: string
          example: "plural_affixes"
        topic_name_tt:
          type: string
          example: "Күплек сан"
        is_public_in_bank:
          type: boolean
          example: false
    CreateTaskRequest:
      type: object
      required:
        - prompt_tt
        - expected_answer
        - cell_count
        - grade_level
        - topic_tag
        - topic_name_tt
      properties:
        prompt_tt:
          type: string
          example: "Сүзгә күплек сан кушымчасын ялгагыз: дус ->"
        expected_answer:
          type: string
          example: "ДУСЛАР"
        cell_count:
          type: integer
          example: 6
        grade_level:
          type: integer
          example: 5
        topic_tag:
          type: string
          example: "plural_affixes"
        topic_name_tt:
          type: string
          example: "Күплек сан"
        is_public_in_bank:
          type: boolean
          example: false
...
    ConstructorTaskRef:
      type: object
      required:
        - task_id
        - order
      properties:
        task_id:
          type: string
          example: "tsk_8201"
        order:
          type: integer
          example: 1
    AssembleTestRequest:
      type: object
      required:
        - title
        - grade_level
        - generate_variants_count
        - task_items
      properties:
        title:
          type: string
          example: "7 сыйныф. 1 нче чирек контроль эше"
        grade_level:
          type: integer
          example: 7
        generate_variants_count:
          type: integer
          example: 2
        task_items:
          type: array
          items:
            $ref: '#/components/schemas/ConstructorTaskRef'
        shuffle_questions_in_variants:
          type: boolean
          example: true
...
    WrongSubmissionItem:
      type: object
      properties:
        student_id:
          type: string
          example: "stu_03"
        student_name:
          type: string
          example: "Хабибуллин Тимур"
        written_answer:
          type: string
          example: "КИТАПГА"
    QuestionAnalyticsItem:
      type: object
      properties:
        question_number:
          type: integer
          example: 1
        marker_id:
          type: integer
          example: 11
        prompt:
          type: string
          example: "Куегыз сүзне юнәлеш килешендә: китап ->"
        expected_answer:
          type: string
          example: "КИТАПКА"
        accuracy_pct:
          type: number
          format: float
          example: 92.0
        wrong_submissions:
          type: array
          items:
            $ref: '#/components/schemas/WrongSubmissionItem'
    AssignmentAnalyticsResponse:
      type: object
      properties:
        assignment_id:
          type: string
          example: "TAT-2026-Q1"
        title:
          type: string
          example: "Исем килешләре һәм кушымчалар"
        total_submissions:
          type: integer
          example: 25
        average_score_pct:
          type: number
          format: float
          example: 83.2
        questions_analytics:
          type: array
          items:
            $ref: '#/components/schemas/QuestionAnalyticsItem'
paths:
...
  /api/v1/assignments/{assignment_id}:
    get:
      tags:
        - Assignment Delivery
      summary: Get Assignment Metadata
      description: Metadata, title, grade level, topic tags, points.
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
          example: "TAT-2026-Q1"
      responses:
        '200':
          description: Metadata, title, grade level, topic tags, points.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentMetadataResponse'
  /api/v1/assignments/{assignment_id}/blank.pdf:
    get:
      tags:
        - Assignment Delivery
      summary: Download Printable Blank PDF (Generic or Single Student)
      description: Download Printable Blank PDF (Generic or Single Student)
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
          example: "TAT-2026-Q1"
        - name: variant
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: Test variant
        - name: student_id
          in: query
          required: false
          schema:
            type: string
          description: Pre-prints student name in header boxes.
      responses:
        '200':
          description: application/pdf binary stream.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...
  /api/v1/assignments/{assignment_id}/offline-bundle:
    get:
      tags:
        - Assignment Delivery
      summary: Download Offline Verification Bundle (Ground-Truth Answer Key)
      description: Downloads everything the mobile app needs to perform 100% offline grading and answer validation on the phone.
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
          example: "TAT-2026-Q1"
        - $ref: '#/components/parameters/HeaderTeacherUUID'
        - name: variant
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfflineBundleResponse'
  /api/v1/constructor/tasks:
    get:
      tags:
        - Assignment Constructor
      summary: Search Task Bank
      description: Search Task Bank
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: topic_tag
          in: query
          required: false
          schema:
            type: string
        - name: grade_level
          in: query
          required: false
          schema:
            type: integer
            minimum: 5
            maximum: 9
          description: 5–9
        - name: max_cells
          in: query
          required: false
          schema:
            type: integer
            maximum: 12
          description: <= 12
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum:
              - all
              - my_tasks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskBankItem'
    post:
      tags:
        - Assignment Constructor
      summary: Create Custom Single Task
      description: Create Custom Single Task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskBankItem'
...
  /api/v1/constructor/tests:
    post:
      tags:
        - Assignment Constructor
      summary: Assemble Modular Test
      description: Composes a new test from selected task IDs with A4 physical limit validation (maximum 8 modular question blocks).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssembleTestRequest'
      responses:
        '201':
          description: Test created
          content:
            application/json:
              schema:
                type: object
                properties:
                  test_id:
                    type: string
                    example: "test_8201_assembled"
                  title:
                    type: string
                    example: "7 сыйныф. 1 нче чирек контроль эше"
                  variants_count:
                    type: integer
                    example: 2
...
  /api/v1/analytics/assignments/{assignment_id}:
    get:
      tags:
        - Backend Analytics Engine
      summary: Assignment Analytics
      description: Item analysis per question across all completed sheets.
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
          example: "TAT-2026-Q1"
        - $ref: '#/components/parameters/HeaderTeacherUUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentAnalyticsResponse'
...